Hackathon/Event_participation_verifier/requirements.txt
streamlit
pdfplumber
openai
python-dotenv
requests
beautifulsoup4
langchain>=0.1.17
langchain-community>=0.0.27
langchain-core>=0.1.50
faiss-cpu
tiktoken
pytesseract
Pillow
PyPDF2